Move ERC20 definitions to api
